Add validation, Id property, and SonarAnalyzer package
Added the `System.Threading.Tasks` namespace to `Department.cs`.
Introduced an `Id` property in the `Department` class as a unique identifier.
Updated the `Name` property in `Department` to use the `required` keyword and added validation attributes for required input and string length constraints.

Added the `SonarAnalyzer.CSharp` package (v10.15.0.120848) to `WebApplicationAPI.csproj` for static code analysis, ensuring it is used only during development.